create notification callback for sending invite and update socket event emitter handleSendInvite
