+++ frontend/src/components/Layout/Modal.tsx
@@ -1,8 +1,12 @@
+/**
+ * Modal that requires user action to accept/decline
+ */
 import React from 'react';
-import { resetModal } from '../../app/features/modalSlice';
-import { User } from '../../app/features/types'
 import { useAppSelector, useAppDispatch } from '../../app/hooks';
+import { User } from '../../app/features/types'
+import { resetModal } from '../../app/features/modalSlice';
 import { handleSendInvite, handleInviteAccepted, handleDeclineInvite } from '../../services/socket/publishers';
+import { setNotificationSendInvite } from '../../util/middleware/socketActions/notification';
 
 const ActionModal: React.FC = () => {
   const dispatch = useAppDispatch();
@@ -14,21 +18,19 @@ const ActionModal: React.FC = () => {
     if (modalData.modalName === 'private chat request') {
       handleDeclineInvite(modalData.peerId!)
     }
+    
     dispatch(resetModal())
   }
 
   const handleAcceptandCloseModal = () => {
     if (modalData.modalName === 'send chat invite') {
-      handleSendInvite(modalData.peerId!, peerUsername)
+      handleSendInvite(modalData.peerId!)
+      setNotificationSendInvite(peerUsername)
     }
 
     if (modalData.modalName === 'private chat request') {
       handleInviteAccepted(modalData.peerId!)
     }
-
-    // if (modalData.modalName === 'start video chat') {
-    //   handleSendVideoInvite()
-    // }
     
     dispatch(resetModal())
   }
